create delegateDirty static prop
