+++ test/phpstan/DynamicFInd.php
@@ -9,6 +9,7 @@
  */
 
 use test\models\Book;
+use function PHPStan\dumpType;
 
 /**
  * Static checking for single model
